backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-docx==0.8.11
requests==2.31.0
python-dotenv==1.0.0
